Move sync-dynamic tests into main test file
